+++ packages/core/src/resolver/resolver-chain-factory.ts
@@ -3,6 +3,7 @@ import {
   EasyDataLoader,
   type MayPromise,
   type Middleware,
+  type RequireKeys,
   applyMiddlewares,
   compose,
   createMemoization,
@@ -15,6 +16,7 @@ import {
   createInputParser,
 } from "./input"
 import {
+  DERIVED_DEPENDENCIES,
   createField,
   createMutation,
   createQuery,
@@ -30,35 +32,70 @@ import type {
   QueryOptions,
 } from "./types"
 
+/**
+ * Interface for chain factory that provides methods to configure GraphQL field options
+ * @template TOutput - The output type of the GraphQL field
+ * @template TInput - The input type of the GraphQL field, can be a single type or a record of types
+ */
 export interface IChainFactory<
   TOutput extends GraphQLSilk,
-  TInput extends
-    | GraphQLSilk
-    | Record<string, GraphQLSilk>
-    | undefined = undefined,
+  TInput extends GraphQLSilk | Record<string, GraphQLSilk> | void = void,
 > {
+  /**
+   * Sets the description for the GraphQL field
+   * @param description - The description text for the field
+   */
   description(description: GraphQLFieldOptions["description"]): this
 
+  /**
+   * Sets the deprecation reason for the GraphQL field
+   * @param deprecationReason - The reason why this field is deprecated
+   */
   deprecationReason(
     deprecationReason: GraphQLFieldOptions["deprecationReason"]
   ): this
 
+  /**
+   * Sets custom extensions for the GraphQL field
+   * @param extensions - Custom extensions to be added to the field
+   */
   extensions(extensions: GraphQLFieldOptions["extensions"]): this
 
+  /**
+   * Sets the output type for the GraphQL field
+   * @template TOutputNew - The new output type
+   * @param output - The output type definition
+   */
   output<TOutputNew extends GraphQLSilk>(
     output: TOutputNew
   ): IChainFactory<TOutputNew, TInput>
 
+  /**
+   * Sets the input type for the GraphQL field
+   * @template TInputNew - The new input type
+   * @param input - The input type definition
+   */
   input<TInputNew extends GraphQLSilk | Record<string, GraphQLSilk>>(
     input: TInputNew
   ): IChainFactory<TOutput, TInputNew>
 }
 
+/**
+ * Options for configuring a chain factory
+ */
 export interface ChainFactoryOptions extends Loom.FieldMeta {
+  /** Middleware functions to be applied to the field */
   middlewares?: Middleware[]
 }
 
+/**
+ * Base class for all chain factories
+ * @template TField - The type of field being created
+ */
 export abstract class BaseChainFactory<TField extends Loom.BaseField = any> {
+  /**
+   * Returns the available methods for the chain factory
+   */
   public static methods() {
     return {
       description: BaseChainFactory.prototype.description,
@@ -67,26 +104,50 @@ export abstract class BaseChainFactory<TField extends Loom.BaseField = any> {
     }
   }
 
+  /**
+   * Creates a new instance of the chain factory
+   * @param options - Configuration options for the factory
+   */
   public constructor(
     protected readonly options?: Partial<ChainFactoryOptions>
   ) {}
 
+  /**
+   * Creates a clone of the current factory with new options
+   * @param options - New options to apply to the clone
+   */
   protected abstract clone(options?: Partial<ChainFactoryOptions>): this
 
+  /**
+   * Sets the description for the field
+   * @param description - The description text
+   */
   public description(description: GraphQLFieldOptions["description"]): this {
     return this.clone({ description })
   }
 
+  /**
+   * Sets the deprecation reason for the field
+   * @param deprecationReason - The reason for deprecation
+   */
   public deprecationReason(
     deprecationReason: GraphQLFieldOptions["deprecationReason"]
   ): this {
     return this.clone({ deprecationReason })
   }
 
+  /**
+   * Sets custom extensions for the field
+   * @param extensions - Custom extensions to add
+   */
   public extensions(extensions: GraphQLFieldOptions["extensions"]): this {
     return this.clone({ extensions })
   }
 
+  /**
+   * Adds middleware functions to the field
+   * @param middlewares - Middleware functions to add
+   */
   public use(...middlewares: Middleware<TField>[]): this {
     return this.clone({
       middlewares: [...(this.options?.middlewares ?? []), ...middlewares],
@@ -94,16 +155,26 @@ export abstract class BaseChainFactory<TField extends Loom.BaseField = any> {
   }
 }
 
+/**
+ * Factory for creating field resolvers with chainable configuration
+ * @template TOutput - The output type of the field
+ * @template TInput - The input type of the field
+ * @template TDependencies - The dependencies of the field
+ */
 export class FieldChainFactory<
     TOutput extends GraphQLSilk = never,
     TInput extends
       | GraphQLSilk
       | Record<string, GraphQLSilk>
       | undefined = undefined,
+    TDependencies extends string[] | undefined = undefined,
   >
-  extends BaseChainFactory<Loom.Field<any, TOutput, TInput>>
+  extends BaseChainFactory<Loom.Field<any, TOutput, TInput, TDependencies>>
   implements IChainFactory<TOutput, TInput>
 {
+  /**
+   * Returns the available methods for the field chain factory
+   */
   public static methods() {
     return {
       ...BaseChainFactory.methods(),
@@ -114,28 +185,63 @@ export class FieldChainFactory<
     } as any as FieldChainFactory<never, undefined>
   }
 
+  /**
+   * Creates a clone of the current factory with new options
+   * @param options - New options to apply to the clone
+   */
   protected clone(options?: Partial<ChainFactoryOptions>): this {
     return new FieldChainFactory({ ...this.options, ...options }) as this
   }
 
+  /**
+   * Sets the output type for the field
+   * @template TOutputNew - The new output type
+   * @param output - The output type definition
+   */
   public output<TOutputNew extends GraphQLSilk>(
     output: TOutputNew
-  ): FieldChainFactory<TOutputNew, TInput> {
+  ): FieldChainFactory<TOutputNew, TInput, TDependencies> {
     return new FieldChainFactory({ ...this.options, output })
   }
 
+  /**
+   * Sets the input type for the field
+   * @template TInputNew - The new input type
+   * @param input - The input type definition
+   */
   public input<TInputNew extends GraphQLSilk | Record<string, GraphQLSilk>>(
     input: TInputNew
-  ): FieldChainFactory<TOutput, TInputNew> {
+  ): FieldChainFactory<TOutput, TInputNew, TDependencies> {
     return new FieldChainFactory({ ...this.options, input })
   }
 
+  /**
+   * Specifies the dependencies for the field
+   * @template TDependencies - The dependencies type
+   * @param dependencies - The dependencies to add
+   */
+  public derivedFrom<const TDependencies extends string[]>(
+    ...dependencies: TDependencies
+  ): FieldChainFactory<TOutput, TInput, TDependencies> {
+    return this.clone({ dependencies }) as any
+  }
+
+  /**
+   * Sets the resolve function for the field
+   * @template TParent - The parent type
+   * @param resolve - The resolve function
+   */
   public resolve<TParent extends GraphQLSilk>(
     resolve: (
-      parent: StandardSchemaV1.InferOutput<TParent>,
+      parent: TDependencies extends string[]
+        ? RequireKeys<
+            NonNullable<StandardSchemaV1.InferOutput<TParent>>,
+            TDependencies[number]
+          >
+        : NonNullable<StandardSchemaV1.InferOutput<TParent>>,
       input: InferInputO<TInput>
     ) => MayPromise<StandardSchemaV1.InferOutput<TOutput>>
-  ): Loom.Field<TParent, TOutput, TInput> {
+  ): Loom.Field<TParent, TOutput, TInput, TDependencies> {
     if (!this.options?.output) throw new Error("Output is required")
     return createField(this.options.output, {
       ...this.options,
@@ -143,12 +249,22 @@ export class FieldChainFactory<
     }) as any
   }
 
+  /**
+   * Sets up data loading for the field
+   * @template TParent - The parent type
+   * @param resolve - The batch resolve function
+   */
   public load<TParent extends GraphQLSilk>(
     resolve: (
-      parents: StandardSchemaV1.InferOutput<TParent>[],
+      parents: (TDependencies extends string[]
+        ? RequireKeys<
+            NonNullable<StandardSchemaV1.InferOutput<TParent>>,
+            TDependencies[number]
+          >
+        : NonNullable<StandardSchemaV1.InferOutput<TParent>>)[],
       input: InferInputO<TInput>
-    ) => MayPromise<StandardSchemaV1.InferOutput<TOutput>[]>
-  ): Loom.Field<TParent, TOutput, TInput> {
+    ) => MayPromise<NonNullable<StandardSchemaV1.InferOutput<TOutput>>[]>
+  ): Loom.Field<TParent, TOutput, TInput, TDependencies> {
     if (!this.options?.output) throw new Error("Output is required")
 
     const useUnifiedParseInput = createMemoization<{
@@ -168,7 +284,9 @@ export class FieldChainFactory<
 
     const operation = "field"
     return meta({
-      ...getFieldOptions(this.options),
+      ...getFieldOptions(this.options, {
+        [DERIVED_DEPENDENCIES]: this.options.dependencies,
+      }),
       operation,
       input: this.options.input as TInput,
       output: this.options.output as TOutput,
@@ -189,20 +307,26 @@ export class FieldChainFactory<
           { parseInput, parent, outputSilk: this.output, operation }
         )
       },
-    }) as Loom.Field<TParent, TOutput, TInput>
+    }) as Loom.Field<TParent, TOutput, TInput, TDependencies>
   }
 }
 
+/**
+ * Factory for creating query resolvers with chainable configuration
+ * @template TOutput - The output type of the query
+ * @template TInput - The input type of the query
+ */
 export class QueryChainFactory<
     TOutput extends GraphQLSilk = never,
-    TInput extends
-      | GraphQLSilk
-      | Record<string, GraphQLSilk>
-      | undefined = undefined,
+    TInput extends GraphQLSilk | Record<string, GraphQLSilk> | void = void,
   >
   extends BaseChainFactory<Loom.Query<TOutput, TInput>>
   implements IChainFactory<TOutput, TInput>
 {
+  /**
+   * Returns the available methods for the query chain factory
+   * @returns An object containing all available methods
+   */
   public static methods() {
     return {
       ...BaseChainFactory.methods(),
@@ -210,25 +334,48 @@ export class QueryChainFactory<
       input: QueryChainFactory.prototype.input,
       resolve: QueryChainFactory.prototype.resolve,
       clone: QueryChainFactory.prototype.clone,
-    } as any as QueryChainFactory<never, undefined>
+    } as any as QueryChainFactory<never, void>
   }
 
+  /**
+   * Creates a clone of the current factory with new options
+   * @param options - New options to apply to the clone
+   * @returns A new instance of QueryChainFactory with the updated options
+   */
   protected clone(options?: Partial<ChainFactoryOptions>): this {
     return new QueryChainFactory({ ...this.options, ...options }) as this
   }
 
+  /**
+   * Sets the output type for the query
+   * @template TOutputNew - The new output type
+   * @param output - The output type definition
+   * @returns A new QueryChainFactory instance with the updated output type
+   */
   public output<TOutputNew extends GraphQLSilk>(
     output: TOutputNew
   ): QueryChainFactory<TOutputNew, TInput> {
     return new QueryChainFactory({ ...this.options, output })
   }
 
+  /**
+   * Sets the input type for the query
+   * @template TInputNew - The new input type
+   * @param input - The input type definition
+   * @returns A new QueryChainFactory instance with the updated input type
+   */
   public input<TInputNew extends GraphQLSilk | Record<string, GraphQLSilk>>(
     input: TInputNew
   ): QueryChainFactory<TOutput, TInputNew> {
     return new QueryChainFactory({ ...this.options, input })
   }
 
+  /**
+   * Sets the resolve function for the query
+   * @param resolve - The resolve function that processes the input and returns the output
+   * @returns A GraphQL query resolver
+   * @throws {Error} If output type is not set
+   */
   public resolve(
     resolve: (
       input: InferInputO<TInput>
@@ -242,6 +389,11 @@ export class QueryChainFactory<
   }
 }
 
+/**
+ * Factory for creating mutation resolvers with chainable configuration
+ * @template TOutput - The output type of the mutation
+ * @template TInput - The input type of the mutation
+ */
 export class MutationChainFactory<
     TOutput extends GraphQLSilk = never,
     TInput extends
@@ -252,6 +404,10 @@ export class MutationChainFactory<
   extends BaseChainFactory<Loom.Mutation<TOutput, TInput>>
   implements IChainFactory<TOutput, TInput>
 {
+  /**
+   * Returns the available methods for the mutation chain factory
+   * @returns An object containing all available methods
+   */
   public static methods() {
     return {
       ...BaseChainFactory.methods(),
@@ -262,22 +418,45 @@ export class MutationChainFactory<
     } as any as MutationChainFactory<never, undefined>
   }
 
+  /**
+   * Creates a clone of the current factory with new options
+   * @param options - New options to apply to the clone
+   * @returns A new instance of MutationChainFactory with the updated options
+   */
   protected clone(options?: Partial<ChainFactoryOptions>): this {
     return new MutationChainFactory({ ...this.options, ...options }) as this
   }
 
+  /**
+   * Sets the output type for the mutation
+   * @template TOutputNew - The new output type
+   * @param output - The output type definition
+   * @returns A new MutationChainFactory instance with the updated output type
+   */
   public output<TOutputNew extends GraphQLSilk>(
     output: TOutputNew
   ): MutationChainFactory<TOutputNew, TInput> {
     return new MutationChainFactory({ ...this.options, output })
   }
 
+  /**
+   * Sets the input type for the mutation
+   * @template TInputNew - The new input type
+   * @param input - The input type definition
+   * @returns A new MutationChainFactory instance with the updated input type
+   */
   public input<TInputNew extends GraphQLSilk | Record<string, GraphQLSilk>>(
     input: TInputNew
   ): MutationChainFactory<TOutput, TInputNew> {
     return new MutationChainFactory({ ...this.options, input })
   }
 
+  /**
+   * Sets the resolve function for the mutation
+   * @param resolve - The resolve function that processes the input and returns the output
+   * @returns A GraphQL mutation resolver
+   * @throws {Error} If output type is not set
+   */
   public resolve(
     resolve: (
       input: InferInputO<TInput>
@@ -291,6 +470,11 @@ export class MutationChainFactory<
   }
 }
 
+/**
+ * Factory for creating subscription resolvers with chainable configuration
+ * @template TOutput - The output type of the subscription
+ * @template TInput - The input type of the subscription
+ */
 export class SubscriptionChainFactory<
     TOutput extends GraphQLSilk = never,
     TInput extends
@@ -301,6 +485,10 @@ export class SubscriptionChainFactory<
   extends BaseChainFactory<Loom.Subscription<TOutput, TInput, any>>
   implements IChainFactory<TOutput, TInput>
 {
+  /**
+   * Returns the available methods for the subscription chain factory
+   * @returns An object containing all available methods
+   */
   public static methods() {
     return {
       ...BaseChainFactory.methods(),
@@ -311,22 +499,46 @@ export class SubscriptionChainFactory<
     } as any as SubscriptionChainFactory<never, undefined>
   }
 
+  /**
+   * Creates a clone of the current factory with new options
+   * @param options - New options to apply to the clone
+   * @returns A new instance of SubscriptionChainFactory with the updated options
+   */
   protected clone(options?: Partial<ChainFactoryOptions>): this {
     return new SubscriptionChainFactory({ ...this.options, ...options }) as this
   }
 
+  /**
+   * Sets the output type for the subscription
+   * @template TOutputNew - The new output type
+   * @param output - The output type definition
+   * @returns A new SubscriptionChainFactory instance with the updated output type
+   */
   public output<TOutputNew extends GraphQLSilk>(
     output: TOutputNew
   ): SubscriptionChainFactory<TOutputNew, TInput> {
     return new SubscriptionChainFactory({ ...this.options, output })
   }
 
+  /**
+   * Sets the input type for the subscription
+   * @template TInputNew - The new input type
+   * @param input - The input type definition
+   * @returns A new SubscriptionChainFactory instance with the updated input type
+   */
   public input<TInputNew extends GraphQLSilk | Record<string, GraphQLSilk>>(
     input: TInputNew
   ): SubscriptionChainFactory<TOutput, TInputNew> {
     return new SubscriptionChainFactory({ ...this.options, input })
   }
 
+  /**
+   * Sets the subscribe function for the subscription
+   * @template TValue - The value type of the subscription
+   * @param subscribe - The subscribe function that returns an AsyncIterator
+   * @returns A subscription resolver that can be further configured with a resolve function
+   * @throws {Error} If output type is not set
+   */
   public subscribe<TValue = StandardSchemaV1.InferOutput<TOutput>>(
     subscribe: (input: InferInputO<TInput>) => MayPromise<AsyncIterator<TValue>>
   ): TValue extends StandardSchemaV1.InferOutput<TOutput>
@@ -357,6 +569,12 @@ export class SubscriptionChainFactory<
   }
 }
 
+/**
+ * Interface for a subscription that can be resolved
+ * @template TOutput - The output type of the subscription
+ * @template TInput - The input type of the subscription
+ * @template TValue - The value type of the subscription
+ */
 export interface ResolvableSubscription<
   TOutput extends GraphQLSilk,
   TInput extends
@@ -365,6 +583,10 @@ export interface ResolvableSubscription<
     | undefined = undefined,
   TValue = StandardSchemaV1.InferOutput<TOutput>,
 > extends Loom.Subscription<TOutput, TInput, TValue> {
+  /**
+   * Sets the resolve function for the subscription
+   * @param resolve - The resolve function
+   */
   resolve(
     resolve: (
       value: TValue,
@@ -373,6 +595,12 @@ export interface ResolvableSubscription<
   ): Loom.Subscription<TOutput, TInput, TValue>
 }
 
+/**
+ * Interface for a subscription that needs to be resolved
+ * @template TOutput - The output type of the subscription
+ * @template TInput - The input type of the subscription
+ * @template TValue - The value type of the subscription
+ */
 export interface SubscriptionNeedResolve<
   TOutput extends GraphQLSilk,
   TInput extends
@@ -381,6 +609,10 @@ export interface SubscriptionNeedResolve<
     | undefined = undefined,
   TValue = StandardSchemaV1.InferOutput<TOutput>,
 > {
+  /**
+   * Sets the resolve function for the subscription
+   * @param resolve - The resolve function
+   */
   resolve(
     resolve: (
       value: TValue,
@@ -464,14 +696,24 @@ export class MutationFactoryWithResolve<
 export class FieldFactoryWithResolve<
   TParent extends GraphQLSilk,
   TOutput extends GraphQLSilk,
-> extends BaseChainFactory<Loom.Field<TParent, TOutput, undefined>> {
-  public get "~meta"(): Loom.Field<TParent, TOutput, undefined>["~meta"] {
+> extends BaseChainFactory<Loom.Field<TParent, TOutput, undefined, undefined>> {
+  public get "~meta"(): Loom.Field<
+    TParent,
+    TOutput,
+    undefined,
+    undefined
+  >["~meta"] {
     return loom.field(this.output, this.options)["~meta"]
   }
 
   public constructor(
     protected output: TOutput,
-    protected readonly options: FieldOptions<TParent, TOutput, undefined>
+    protected readonly options: FieldOptions<
+      TParent,
+      TOutput,
+      undefined,
+      undefined
+    >
   ) {
     super(options)
   }